+++ android-app/app/src/test/java/com/lorenzo/mangadownloader/HomeScreenUiTest.kt
@@ -30,6 +30,7 @@ class HomeScreenUiTest {
                     editMode = false,
                     padding = PaddingValues(),
                     onResume = {},
+                    onResumeStreamingChapter = {},
                     onOpenUpdate = {},
                     onOpenAllUpdates = {},
                     onOpenHistory = {},